backend/chroma_api/requirements.txt
fastapi
uvicorn
chromadb
langchain-openai
python-dotenv